fix printf formating

+++ internal/api/test_helper_common.h
@@ -47,13 +47,13 @@ static inline int test_validate_status(int ret, uint64_t algorithm)
 		ret += 1;
 #endif
 		printf("Self test status for algorithm %llu unexpected: %u\n",
-		       algorithm, lc_status_get_result(algorithm));
+		       (unsigned long long)algorithm,
+		       lc_status_get_result(algorithm));
 	}
 #ifdef LC_FIPS140_DEBUG
 	else {
-		printf("FIPS negative test: Self test status for algorithm %" PRIu64
-		       " failed as expected\n",
-		       algorithm);
+		printf("FIPS negative test: Self test status for algorithm %llu failed as expected\n",
+		       (unsigned long long)algorithm);
 	}
 #endif
